requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0
eventlet==0.33.3
lxml==4.9.3
selenium==4.15.2
schedule==1.2.0
redis==4.5.4